+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # main [(unreleased)](https://github.com/fastruby/dotenv_validator/compare/v1.0.0...main)
 
 * [BUGFIX] Fix the way that we calculate the path to the .env.sample file (by [@etagwerker][])
+* [Feature] Add boolean format
 
 # v1.0.0 / 2021-07-29 [(commits)](https://github.com/fastruby/dotenv_validator/tree/v1.0.0)